pr comment - moved apierror to model
